Started new getAverageStats function to calculate average stats for a character based on their prior generations. This function takes in the character ID and the stats from generations 0-3, combines them, and divides by 4 to get the average. It also ensures that the average stats do not exceed the maximum stats for that character. Changed int to double for StatsPrediction class.

+++ feth_character_planner/lib/models/plan.dart
@@ -95,9 +95,9 @@ class ClassHistory {
 
 // Stats Prediction Start
 class StatsPrediction {
-  Map<String, int> rng1;
-  Map<String, int> rng2;
-  Map<String, int> rng3;
+  Map<String, double> rng1;
+  Map<String, double> rng2;
+  Map<String, double> rng3;
 
   StatsPrediction({
     required this.rng1,
@@ -126,7 +126,7 @@ class StatsPrediction {
   }
 
 
-  static Map<String,int> _convertStats(
+  static Map<String,double> _convertStats(
       Map<String,dynamic> stats
   ){
